+++ src/main/java/it/gov/pagopa/pu/registry/model/SendTimelineRegistry.java
@@ -44,7 +44,7 @@ public class SendTimelineRegistry extends BaseEntity {
   @NotNull
   private String notificationRequestId;
   private String iun;
-  private Long recipientIndex;
+  private Integer recipientIndex;
   private String newStatus;
 
   @NotNull
